backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==0.27.0
pydantic==1.8.2
whisper==1.0.0
requests==2.26.0